updates to formatting
The printed values are now rounded up to the last three digits. The choice of using a ceiling function over Python's round-function was made because the latter performed worse. The downside is that the values my now be off by 1 in the last digit.
In addition to that, the spacing when printing the policies is now better but still not perfect. Proper formatting seems not to be possible since unicode symbols cannot be formatted with string formatting tools.